+++ src/js/src/index.js
@@ -1,8 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import htm from "htm";
-
-const html = htm.bind(React.createElement);
 
 export function bind(node) {
   return {
@@ -15,30 +12,69 @@ export function bind(node) {
   };
 }
 
-export function History({ onChange }) {
-  // capture changes to the browser's history
+export function History({ onHistoryChange }) {
+  // Capture browser "history go back" action and tell the server about it
+  // Note: Browsers do not allow us to detect "history go forward" actions.
   React.useEffect(() => {
+    // Register a listener for the "popstate" event and send data back to the server using the `onHistoryChange` callback.
     const listener = () => {
-      onChange({
+      onHistoryChange({
         pathname: window.location.pathname,
         search: window.location.search,
       });
     };
+
+    // Register the event listener
     window.addEventListener("popstate", listener);
+
+    // Delete the event listener when the component is unmounted
     return () => window.removeEventListener("popstate", listener);
   });
-  return null;
-}
 
-export function Link({ to, onClick, children, ...props }) {
-  const handleClick = (event) => {
-    event.preventDefault();
-    window.history.pushState({}, to, new URL(to, window.location));
-    onClick({
+  // Tell the server about the URL during the initial page load
+  // FIXME: This currently runs every time any component is mounted due to a ReactPy core rendering bug.
+  // https://github.com/reactive-python/reactpy/pull/1224
+  React.useEffect(() => {
+    onHistoryChange({
       pathname: window.location.pathname,
       search: window.location.search,
     });
-  };
+    return () => {};
+  }, []);
+  return null;
+}
 
-  return html`<a href=${to} onClick=${handleClick} ...${props}>${children}</a>`;
+// FIXME: The Link component is unused due to a ReactPy core rendering bug
+// which causes duplicate rendering (and thus duplicate event listeners).
+// https://github.com/reactive-python/reactpy/pull/1224
+export function Link({ onClick, linkClass }) {
+  // This component is not the actual anchor link.
+  // It is an event listener for the link component created by ReactPy.
+  React.useEffect(() => {
+    // Event function that will tell the server about clicks
+    const handleClick = (event) => {
+      event.preventDefault();
+      let to = event.target.getAttribute("href");
+      window.history.pushState({}, to, new URL(to, window.location));
+      onClick({
+        pathname: window.location.pathname,
+        search: window.location.search,
+      });
+    };
+
+    // Register the event listener
+    let link = document.querySelector(`.${linkClass}`);
+    if (link) {
+      link.addEventListener("click", handleClick);
+    }
+
+    // Delete the event listener when the component is unmounted
+    return () => {
+      let link = document.querySelector(`.${linkClass}`);
+      if (link) {
+        link.removeEventListener("click", handleClick);
+      }
+    };
+  });
+  return null;
 }