+++ src/reactpy_router/__init__.py
@@ -1,20 +1,16 @@
 # the version is statically loaded by setup.py
 __version__ = "0.1.1"
 
-from . import simple
-from .core import create_router, link, route, router_component, use_params, use_query
-from .types import Route, RouteCompiler, RouteResolver
+
+from .components import link, route
+from .hooks import use_params, use_search_params
+from .routers import browser_router, create_router
 
 __all__ = (
     "create_router",
     "link",
     "route",
-    "route",
-    "Route",
-    "RouteCompiler",
-    "router_component",
-    "RouteResolver",
-    "simple",
+    "browser_router",
     "use_params",
-    "use_query",
+    "use_search_params",
 )