+++ tests/conftest.py
@@ -1,31 +1,44 @@
+import asyncio
+import os
+import sys
+
 import pytest
 from playwright.async_api import async_playwright
 from reactpy.testing import BackendFixture, DisplayFixture
 
+GITHUB_ACTIONS = os.getenv("GITHUB_ACTIONS", "").lower() == "true"
+
 
 def pytest_addoption(parser) -> None:
     parser.addoption(
-        "--headed",
-        dest="headed",
+        "--headless",
+        dest="headless",
         action="store_true",
-        help="Open a browser window when runnging web-based tests",
+        help="Hide the browser window when running web-based tests",
     )
 
 
 @pytest.fixture
 async def display(backend, browser):
-    async with DisplayFixture(backend, browser) as display:
-        display.page.set_default_timeout(10000)
-        yield display
+    async with DisplayFixture(backend, browser) as display_fixture:
+        display_fixture.page.set_default_timeout(10000)
+        yield display_fixture
 
 
 @pytest.fixture
 async def backend():
-    async with BackendFixture() as backend:
-        yield backend
+    async with BackendFixture() as backend_fixture:
+        yield backend_fixture
 
 
 @pytest.fixture
 async def browser(pytestconfig):
     async with async_playwright() as pw:
-        yield await pw.chromium.launch(headless=not bool(pytestconfig.option.headed))
+        yield await pw.chromium.launch(headless=True if GITHUB_ACTIONS else pytestconfig.getoption("headless"))
+
+
+@pytest.fixture
+def event_loop_policy(request):
+    if sys.platform == "win32":
+        return asyncio.WindowsProactorEventLoopPolicy()
+    return asyncio.get_event_loop_policy()